Implement Identity.getProfile().
